backend/requirements.txt
# FastAPI Framework and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# Data Validation and Serialization
pydantic[email]==2.5.0

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Password Hashing
argon2-cffi==23.1.0

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Image Processing and File Handling
Pillow==10.1.0

# Payment Processing
stripe==7.8.0

# Email Validation (used by pydantic EmailStr)
email-validator==2.1.0

# HTTP Requests (for SMS services and external APIs)  
requests==2.31.0
httpx==0.25.2

# SMS Services (optional - uncomment if using)
# twilio==8.10.0  # For Twilio SMS

# Date and Time Utilities
python-dateutil==2.8.2

# CORS Middleware (included with FastAPI but explicit)
# Note: CORS is included with FastAPI, but listed for clarity

# Development and Testing (optional - uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-mock==3.12.0

# Logging and Monitoring (optional - uncomment if needed)
# loguru==0.7.2

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1

# For handling ObjectIds and MongoDB operations
# Note: pymongo includes bson, but listing for clarity